Pulled activation 'definitions' to separate folder
